requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.7.1
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==25.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
redis==7.0.1
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
zope.interface==8.1.1
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
